feat(slider)!: add a groupSeparator prop to enable toggling display of separator in label (#5138)
**Related Issue:** #4976

## Summary

BREAKING CHANGE: For consistency, a `groupSeparator` prop was introduced to opt-into including the localized, numeric group separator in labels. Before this change, labels automatically included the group separator when rendered. In order to keep this behavior, developers will need to set `groupSeparator = true` (prop), or `group-separator` (attribute)  on the component. 

Co-authored-by: JC Franco <jfranco@esri.com>