ci(screener): refine when screener runs on pull requests (#5280)
**Related Issue:** #5253

## Summary
This PR refines the PR linked above. Here are the conditions for when
Screener runs on a PR:

- Screener will run when when the `pr ready for visual snapshots` label
is added to a normal PR (not draft PRs)
- Screener will run when a draft PR with the `pr ready for visual
snapshots` label is marked as "Ready to review"

When someone pushes changes to a PR that has the `pr ready for visual
snapshots label` (meaning snapshots already ran), __Screener will NOT
run but it WILL show a green checkmark__ allowing the author to merge
the PR. This is so that PRs can be merged w/o another Screener run if
there are some small cleanups needed.

If you need to rerun screener, remove and re-add the `pr ready for
visual snapshots` label.

Screener will still run on every push to `master`. This should provide a
back up in case a visual regression slips through due to the point
above. I will monitor the Screener runs on `master` while these
limitations are in place.
<!--

Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->