feat(pagination)!: add groupSeparator prop to enable toggling display of separator in pages (#5265)
**Related Issue:** #5259

BREAKING CHANGE: For consistency, a groupSeparator prop was introduced to opt-into including the localized, numeric group separator in page numbers. Before this change, page numbers automatically included the group separator when rendered. In order to keep this behavior, developers will need to set groupSeparator = true(prop), or group-separator (attribute)  on the component.