ci: changelog modifications aren't overwritten after next releases (#5791)
**Related Issue:** #

## Summary
It turns out the default behavior of conventional-changelog is to
generate the changelog items since the most recent tag. Whoever wrote
the script was deleting all of the next tags and readding them so that
the whole changelog would be regenerated. I ended up writing some
helpers to combine the existing unreleased changelog items and the new
ones.