feat(input, input-text, input-number): add attributes autocomplete, accept, multiple, pattern (#5807)
**Related Issue:** #4079 

## Summary
Adds native input attributes:

-
[multiple](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/multiple)
to `calcite-input` for the types specified on MDN
-
[accept](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/accept)
to `calcite-input` for the types specified on MDN
-
[pattern](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/pattern)
to `calcite-input-text` and `calcite-input` for the types specified on
MDN
-
[autocomplete](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete)
to `calcite-input-number`, `calcite-input-text`, and `calcite-input` for
the types specified on MDN

## Note
just saw on MDN after committing that Safari may not support
`autocomplete`:

https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/autocomplete#browser_compatibility

Should I remove it from the PR?


<!--
Please make sure the PR title and/or commit message adheres to the
https://www.conventionalcommits.org/en/v1.0.0/ specification.

Note: If your PR only has one commit and it is NOT semantic, you will
need to either

a. add another commit and wait for the check to update
b. proceed to squash merge, but make sure the commit message is the same
as the title.

This is because of the way GitHub handles single-commit squash merges
(see https://github.com/zeke/semantic-pull-requests/issues/17)

If this is component-related, please verify that:

- [ ] feature or fix has a corresponding test
- [ ] changes have been tested with demo page in Edge

---

If this is skipping an unstable test:

- include info about the test failure
- submit an unstable-test issue by
[choosing](https://github.com/Esri/calcite-components/issues/new/choose)
the unstable test template and filling it out

-->