fix(tooltip): Fix hover logic for elements within shadowRoot (#6119)
**Related Issue:** #6113

## Summary

fix(tooltip): Fix hover logic for elements within shadowRoot. #6113

It looks like the events `pointerover` and `pointerout` won't be
triggered on elements within a shadowRoot.

The alternative this PR uses it to switch to using just `pointermove` so
that It can query for a tooltip anytime the pointer moves. Because this
can cause a lot of events to occur, this PR also moves `setTimeout`
logic on `pointermove` so there is no performance decrease.