fix(input, input-number): correctly handle '-' and '.' values when sanitizing number (#6306)
**Related Issue:** #6270

## Summary
Fixes issues in input and input-number involving how to handle `-` and
`.` values. We don't want to set it as the `value` property because they
aren't numbers, but we also don't want to sanitize them because people
should be allowed to type negative and decimal numbers. Added a couple
code comments for clarity

Co-authored-by: JC Franco <jfranco@esri.com>