ci(next): fix commit message (#6425)
**Related Issue:** #6406

## Summary

The commit messages for `next` releases are goofed. This removes our
custom `{{currentTag}}` commit message and replaces it with the default:
`chore(release): {{currentTag}}`

It also removes `noVerify` so we can remove `npm run lint:md` from 
standard-version's pre-commit hook because it will be done by `lint-staged`.

The commit message was fine before the PR linked above. No idea why this
is happening but it's probably a `standard-version` bug.