fix: restore form control validation in Safari (#6623)
**Related Issue:** #6626

## Summary

This restores consistent usage of `requestSubmit` when submitting forms.
`submit` and `requestSubmit` [behave
differently](https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/requestSubmit#usage_notes)
and it was currently not going through form validation in Safari. This
moves the polyfill to the form util to ensure it's available for
components using the form util.