feat(combobox): make combobox clearable (#6972)
**Related Issue:** #4738 

## Summary

This updates the combobox to display a clear-value button by default and
adds a `clearDisabled` property to disable the clear-button.
Additionally, this prevents single-selection from deselecting a value
when clicked again making it similar to `calcite-select`.

## Noteworthy changes

* Includes a functional component + style mixin to create an internal
close or x button used in other places. This can be used as the base for
an internal x-button component if needed.
* Drops the `coerce` param from `combobox-item#toggleSelected` as it's
internal and isn't used.