feat(tab-nav): adds optional closable functionality to individual `tab-titles` (#6740)
**Related Issue:** #2620

## Summary
`handleTabTitleClose` sequence is as follows:
> if `closed` item is `selected`, fall back on the next
> unless it's the last item, which falls back on previous
> closing non-selected items doesn't affect the `selected` tab
> last remaining item becomes `disabled`, if `closable`

- Added `closed` property
- Added `closable` property
- Added `calciteInternalTabsClose` event
- Added `calciteTabsClose` event
- Added t9n messages for `Close` text.
- Updated styling per new Figma specs
- Added tests

`handleTabTitleClose` keeps track of `visibleTabTitlesIndices` instead
of having Dom queries filter out `hidden`. This helps to sync `tabs`
with `tab-titles` using their corresponding `id`s set on the original
render. `visibleTabIndeces` is an array of `id`s of the non-hidden
`tab-titles`. When `tab-title` is `closed` this `id` is used to navigate
to the next visible item. This makes figuring out what the next item
should be when closing out of order is now more straightforward.

Tests coverage for 

- simple closing behavior 
- closing sequence behavior:
> - when closing `tab-titles` in sequence 1 (first selected) through 4,
`tab-title` and corresponding `tab` become hidden, and selection
fallback is the next tab
> - reverse fallback: when closing `tab-titles` in sequence 4 (last
selected) through 1, `tab-title` and corresponding `tab` become hidden,
and selection fallback is the previous tab
> - closing an unselected `tab-title` does not deselect the current
selection
> - 2 additional cases: works with a randomized closing sequence with
mixed selected and not

---------

Co-authored-by: Matt Driscoll <mdriscoll@esri.com>
Co-authored-by: JC Franco <jfranco@esri.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Anveshreddy mekala <anv11827@esri.com>
Co-authored-by: jona7150 <jturpin@esri.com>
Co-authored-by: Kitty Hurley <khurley@esri.com>