style: lint and format code (#7319)
**Related Issue:** #7318

## Summary

Lint and format the code. Prettier will now add trailing commas when
possible, which is the default in 3x.

I know already brought up disabling some jsdoc rules in the past, but I
think we can at least disable the ` jsdoc/require-param-type` one. We
are using TypeScript so defining the types twice seems redundant and
likely to get out of sync when things change.