ci: notify teams about calcite-components releases only (#7317)
**Original PRs:** #3771 #5512

## Summary

Adding back the Teams release notification action. I had to refactor it
after the monorepo transition so it wouldn't spam the channel with
release notifications for all of the packages.

We can make additional steps for different packages that go to different
channels at some point. For example, a `design-tokens` release
notification to "Get Help - Design and FIGMA"