fix(color-picker): maintains correct numbering system when entering invalid RGB value (#7327)
**Related Issue:** #5978

## Summary

Entering an RGB channel value greater than 255 or less than 0 clamps the
value so it doesn't become invalid. This caused the numbering system to
revert to the default `latn`.

It is a timing issue with `calcite-input`'s lifecycle methods and it
reproduces outside of `calcite-color-picker`. Here is a demo showing
that using a timeout or requesting an animation frame (kind of) fixes
the issue:

https://codepen.io/benesri/pen/vYQjYJK

The issue isn't reproducible in `calcite-input-number`.