ci: ensure vscode-data is generated during every build (#7480)
**Related Issue:** #7397 

## Summary

The vscode-data Stencil output is not being generated if one already
exists. It works on `next` because the components get built 3 times,
once for tests, once for storybook deployment, and then a third time for
the stencil types workaround. I'll try to find a way to remove one of
those builds for `next` deployments.