refactor(tabs, tab, tab-nav, tab-title): `getElementProp` is refactored out in favor of inheritable props set directly on parent (#7605)
**Related Issue:** #6038

## Summary

`getElementProp` is refactored out across child components as an
outdated pattern in favor of inheritable props set directly on the
parent. Instead of the `tab, tab-title, tab-nav` looking up the parent
for `position` and `scale`, these get set by the `tabs` parent.

The logic for setting these props thus moves to the parent, getting rid
of the `getElementProp` altogether. The parent component gets a
`mutationObserver` to achieve this and watchers for when it needs to
modify the children. Child attribute selectors are subbed with new
classes added dynamically to reflect the scale and position, as
attributes are no longer reflected to be targeted.


Inherited props addressed:

- [x] position
- [x] scale