build(angular): chage calcite-components to a dep from peerDep/devDep (#8167)
Changes calcite-components to a dependency of the angular wrapper,
versus a peerDep and devDep.

Stencil has it as a peerDep because it is required for the user.
arcgis-web-components and us added it as a devDep so the monorepo
dependency tree would catch it.

Putting it as a dep resolves both of the above, and it prevents the user
from needing to install it on their end. It seems like lerna doesn't
bump the peerDep when versioning local packages so it falls behind,
which causes issues.

Note: we did the same thing for the react wrapper for convenience, even
before we moved it to the monorepo