ci(next): attempt to fix tests by copying steps in passing pr check (#8179)
## Summary

We have consistent timeout test failures on `main`, which is preventing
`next` deployments. Strangely, the tests are passing on the PR check and
locally. I'm not sure what's going on and it's hard to troubleshoot
because it only occurs in the CI.

I'm adding a "supposedly" redundant build to the next deployment
workflow to align with the passing e2e PR check:


https://github.com/Esri/calcite-design-system/blob/91b7c0b65950bb624b4c8a09132d9095510e178e/.github/workflows/pr-e2e.yml#L32

I must have forgotten to remove the extra build in the PR check when I
set up the monorepo. It shouldn't be required because we have turbo set
up to build before testing:


https://github.com/Esri/calcite-design-system/blob/91b7c0b65950bb624b4c8a09132d9095510e178e/turbo.json#L13-L15

This shouldn't add additional time because turbo will use the cache from
the previous build during `npm test`. Maybe a second or two for cache
lookup, but nothing substantial.

If this hail mary doesn't work (which is my guess), then it's back to
the drawing board. If it does work somehow, then additional
investigation/changes will be required on my end to untangle things.