fix(react)!: disable includeImportCustomElements to resolve initial render issues (#8248)
**Related Issue:** #8143

## Summary

The `ReferenceError: navigator is not defined` error is still present in
Stencil v4 when removing the patch. Disabled
[`includeImportCustomElements`](https://stenciljs.com/docs/react#includeimportcustomelements)
for now while we investigate an alternative fix that doesn't cause
initial render issues.


BREAKING CHANGE: Disabled `includeImportCustomElements`. Make sure to
import components from `@esri/calcite-components` in addition to the
react wrappers. For example, the first code snippet in #7185 is now
required, or else the custom elements will not be defined on the window.